moves landing page graphs into landing page folder
